custom_components/embedded_google_assistant/requirements.txt
google-assistant-grpc==0.2.1
google-auth-oauthlib==0.4.1
# sounddevice>=0.3.7,<0.4